+++ packages/fuels/src/cli/commands/dev/autoStartFuelCore.test.ts
@@ -1,39 +1,101 @@
-import { fuelsConfig } from '../../../../test/fixtures/fuels.config';
+import * as testUtilsMod from '@fuel-ts/wallet/test-utils';
+import { existsSync, rmSync } from 'fs';
+import { join } from 'path';
 
+import { fuelsConfig } from '../../../../test/fixtures/fuels.config';
+import type { FuelsConfig } from '../../types';
+
+import type { FuelCoreNode } from './autoStartFuelCore';
 import { autoStartFuelCore } from './autoStartFuelCore';
-import * as startFuelCoreMod from './startFuelCore';
+
+vi.mock('@fuel-ts/wallet/test-utils', async () => {
+  const mod = await vi.importActual('@fuel-ts/wallet/test-utils');
+  return {
+    __esModule: true,
+    ...mod,
+  };
+});
 
 /**
  * @group node
  */
 describe('autoStartFuelCore', () => {
-  function mockStartFuelCore() {
-    const startFuelCore = vi
-      .spyOn(startFuelCoreMod, 'startFuelCore')
-      .mockResolvedValue({} as unknown as startFuelCoreMod.FuelCoreNode);
+  afterEach(() => {
+    vi.restoreAllMocks();
 
-    return { startFuelCore };
+    const chainConfig = join(fuelsConfig.basePath, '.fuels', 'chainConfig.json');
+    if (existsSync(chainConfig)) {
+      rmSync(chainConfig);
+    }
+  });
+
+  function mockLaunchNode() {
+    vi.spyOn(testUtilsMod, 'launchNode').mockImplementation((..._: unknown[]) =>
+      Promise.resolve({ cleanup: () => {}, ip: '0.0.0.0', port: '4000' })
+    );
   }
 
   test('should auto start `fuel-core`', async () => {
-    const { startFuelCore } = mockStartFuelCore();
+    mockLaunchNode();
 
     const config = structuredClone(fuelsConfig);
     config.autoStartFuelCore = true;
 
     await autoStartFuelCore(config);
 
-    expect(startFuelCore).toHaveBeenCalledTimes(1);
+    expect(testUtilsMod.launchNode).toHaveBeenCalledTimes(1);
   });
 
   test('should not start `fuel-core`', async () => {
-    const { startFuelCore } = mockStartFuelCore();
+    mockLaunchNode();
 
     const config = structuredClone(fuelsConfig);
     config.autoStartFuelCore = false;
 
     await autoStartFuelCore(config);
 
-    expect(startFuelCore).toHaveBeenCalledTimes(0);
+    expect(testUtilsMod.launchNode).toHaveBeenCalledTimes(0);
+  });
+
+  test('should start `fuel-core` node using built-in binary', async () => {
+    mockLaunchNode();
+
+    const copyConfig: FuelsConfig = structuredClone(fuelsConfig);
+    copyConfig.useBuiltinFuelCore = true;
+
+    // this will cause it to autofind a free port
+    copyConfig.fuelCorePort = undefined;
+    delete copyConfig.fuelCorePort;
+
+    const core = (await autoStartFuelCore(copyConfig)) as FuelCoreNode;
+
+    expect(testUtilsMod.launchNode).toHaveBeenCalledTimes(1);
+
+    expect(core.bindIp).toEqual('0.0.0.0');
+    expect(core.accessIp).toEqual('127.0.0.1');
+    expect(core.port).toBeGreaterThanOrEqual(4000);
+    expect(core.providerUrl).toMatch(/http:\/\/127\.0\.0\.1:([0-9]+)\/graphql/);
+    expect(core.killChildProcess).toBeTruthy();
+
+    core.killChildProcess();
+  });
+
+  test('should start `fuel-core` node using system binary', async () => {
+    mockLaunchNode();
+
+    const core = (await autoStartFuelCore({
+      ...structuredClone(fuelsConfig),
+      useBuiltinFuelCore: false,
+    })) as FuelCoreNode;
+
+    expect(testUtilsMod.launchNode).toHaveBeenCalledTimes(1);
+
+    expect(core.bindIp).toEqual('0.0.0.0');
+    expect(core.accessIp).toEqual('127.0.0.1');
+    expect(core.port).toBeGreaterThanOrEqual(4000);
+    expect(core.providerUrl).toMatch(/http:\/\/127\.0\.0\.1:([0-9]+)\/graphql/);
+    expect(core.killChildProcess).toBeTruthy();
+
+    core.killChildProcess();
   });
 });
