+++ packages/fuels/src/cli/commands/dev/index.ts
@@ -9,8 +9,8 @@ import { build } from '../build';
 import { deploy } from '../deploy';
 import { withConfigErrorHandler } from '../withConfig';
 
+import type { FuelCoreNode } from './autoStartFuelCore';
 import { autoStartFuelCore } from './autoStartFuelCore';
-import type { FuelCoreNode } from './startFuelCore';
 
 export const closeAllFileHandlers = (handlers: FSWatcher[]) => {
   handlers.forEach((h) => h.close());
